+++ src/GRA.Domain.Service/ActivityService.cs
@@ -787,7 +787,8 @@ namespace GRA.Domain.Service
 
                 if (trigger.AwardAvatarBundleId.HasValue)
                 {
-                    await AwardUserBundle(userId, trigger.AwardAvatarBundleId.Value);
+                    await AwardUserBundle(userId, trigger.AwardAvatarBundleId.Value,
+                        userIdIsCurrentUser);
                 }
 
                 // send mail if applicable
@@ -1098,11 +1099,13 @@ namespace GRA.Domain.Service
             }
         }
 
-        private async Task AwardUserBundle(int userId, int bundleId)
+        private async Task AwardUserBundle(int userId, int bundleId,
+            bool userIdIsCurrentUser = false)
         {
             var bundle = await _dynamicAvatarBundleRepository.GetByIdAsync(bundleId);
             if (bundle.DynamicAvatarItems.Count > 0)
             {
+                var loggingUser = (userIdIsCurrentUser ? userId : GetActiveUserId());
                 var userItems = await _dynamicAvatarItemRepository.GetUserUnlockedItemsAsync(userId);
 
                 var newItems = bundle.DynamicAvatarItems.Select(_ => _.Id).Except(userItems).ToList();
@@ -1125,9 +1128,9 @@ namespace GRA.Domain.Service
                     notification.Text += " You can view the full list of pieces unlocked in your Profile History.";
                 }
 
-                await _notificationRepository.AddSaveAsync(userId, notification);
+                await _notificationRepository.AddSaveAsync(loggingUser, notification);
 
-                await _userLogRepository.AddSaveAsync(GetActiveUserId(), new UserLog
+                await _userLogRepository.AddSaveAsync(loggingUser, new UserLog
                 {
                     UserId = userId,
                     PointsEarned = 0,