+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Fixed
 - Issue where school CSV import fails if two schools have the same name
 - Fix secret code GraException message typo
+- Fix bundle award on login
 
 ## [4.0.0-alpha6] 2017-06-02
 ### Added